api/requirements/base.txt
Django==5.0.14
python-dotenv==1.0.1
djangorestframework==3.15.0
django-autoslug==1.9.9
django-countries==7.5.1
django-phonenumber-field==8.3.0
phonenumbers==8.13.32
drf-yasg==1.21.7
Pillow==11.3.0
argon2-cffi==23.1.0
djoser==2.2.2
django-taggit==5.0.1
django-filter==24.1
django-celery-email==3.0.0
cloudinary==1.39.1
django-celery-beat==2.6.0
redis==5.0.3
celery==5.3.6
flower==2.0.1
